Broke down the components
