+++ src/app/components/productListing/filterSorting/sortBySection.tsx
@@ -1,0 +1,43 @@
+import { useDispatch, useSelector } from "react-redux";
+import Accordion from "../../accordion/accordion";
+import { RootState } from "@/app/stores";
+import { sorts } from "@/data/filters";
+import { setSorting } from "@/app/stores/sortingSlice";
+
+export const SortBySection = () => {
+  const dispatch = useDispatch();
+
+  const selectedSorting = useSelector(
+    (state: RootState) => state.sorting.selectedSorting
+  );
+
+  const handleSorting = (sortingOption: string) => {
+    dispatch(setSorting(sortingOption));
+  };
+
+  return (
+    <Accordion
+      title="SORT BY"
+      outerSectionClassName="pb-0"
+      sectionClassName="mx-6"
+    >
+      {sorts.map((sort, index) => (
+        <div
+          key={index}
+          className={`w-full ${
+            index < sorts.length - 1 ? "border-b border-gray-200" : ""
+          }`}
+          onClick={() => handleSorting(sort.id)}
+        >
+          <div
+            className={`flex flex-row w-full hover:cursor-pointer hover:bg-gray-200 transition-all duration-300 ${
+              selectedSorting === sort.id ? "border-l-4 border-black" : ""
+            }`}
+          >
+            <span className="mx-6 my-3 uppercase">{sort.name}</span>
+          </div>
+        </div>
+      ))}
+    </Accordion>
+  );
+};
